refactor: 컴포넌트 구분 및 ui 디자인 적용
